Cleanup, revert older version to original
