--- weather@mockturtl/.babelrc
@@ -1,9 +1,0 @@
-{
-  "presets": [
-    ["@babel/preset-env", {
-      "targets": {
-        "chrome": "58",
-        "ie": "11"
-      }
-  }]],
-}
