--- weather@mockturtl/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "weather-mockturtl",
-  "version": "1.0.0",
-  "description": "Adaptation of Gnome Shell's [weather extension](https://github.com/simon04/gnome-shell-extension-weather) for the [Cinnamon](http://cinnamon.linuxmint.com) desktop.",
-  "main": "files/weather@mockturtl/3.8/applet.js",
-  "scripts": {
-    "build": "./node_modules/.bin/babel src -d files/weather@mockturtl/3.6",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@babel/cli": "^7.2.3",
-    "@babel/core": "^7.4.0",
-    "@babel/preset-env": "^7.4.2"
-  },
-  "dependencies": {
-    "@babel/plugin-transform-runtime": "^7.4.0",
-    "regenerator-runtime": "^0.13.2"
-  }
-}
